backend/requirements.txt
Django==6.0.3
Pillow==12.1.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
google-generativeai==0.8.5
python-dotenv==1.0.1
